+++ frontend/app/src/router/routes.ts
@@ -52,6 +52,7 @@ export const Routes = {
   SETTINGS_ACCOUNT: ensureRoute('/settings/account'),
   SETTINGS_ACCOUNTING: ensureRoute('/settings/accounting'),
   SETTINGS_DATABASE: ensureRoute('/settings/database'),
+  SETTINGS_EVM: ensureRoute('/settings/evm'),
   SETTINGS_GENERAL: ensureRoute('/settings/general'),
   SETTINGS_INTERFACE: ensureRoute('/settings/interface'),
   SETTINGS_MODULES: ensureRoute('/settings/modules'),
@@ -282,6 +283,11 @@ export const useAppRoutes = createSharedComposable(() => {
       icon: 'lu-manage-assets' as const,
       route: Routes.SETTINGS_DATABASE,
       text: t('navigation_menu.settings_sub.database'),
+    },
+    SETTINGS_EVM: {
+      icon: 'lu-link' as const,
+      route: Routes.SETTINGS_EVM,
+      text: t('navigation_menu.settings_sub.evm'),
     },
     SETTINGS_GENERAL: {
       icon: 'lu-bolt' as const,