+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`` Users will now be able to change the order of used indexers for EVM chains.
 * :bug:`-` Curve pool native asset withdrawals will now include the pool in the event notes like withdrawals of other tokens.
 * :feature:`-` rotki will now properly decode Yearn staking deposit, withdrawal, and reward claim transactions.
 * :feature:`11030` rotki will now support cowswap transactions on Polygon PoS.